+++ artemis-backend-gwt/artemis-gwt/src/main/java/com/artemis/gwtref/client/IReflectionCache2.java
@@ -18,23 +18,5 @@ package com.artemis.gwtref.client;
 
 import java.util.Collection;
 
-public interface IReflectionCache {
-	// Class level methods
-	public Type forName (String name);
-
-	public Object newArray (Type componentType, int size);
-
-	public int getArrayLength (Type type, Object obj);
-
-	public Object getArrayElement (Type type, Object obj, int i);
-
-	public void setArrayElement (Type type, Object obj, int i, Object value);
-
-	// Field Methods
-	public Object get (Field field, Object obj) throws IllegalAccessException;
-
-	public void set (Field field, Object obj, Object value) throws IllegalAccessException;
-
-	// Method Methods :p
-	public Object invoke (Method m, Object obj, Object[] params);
+public interface IReflectionCache2 extends IReflectionCache {
 }