+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Change Log (we do our best to adhere to [semantic versioning](http://semver.org/))
 
+#### Version: 2.5.0-SNAPSHOT
+
+- **BREAKING CHANGES**
+   - Minor changes to reflection API to match libGDX 1.10 API.  
+       ClassReflection::isAbstractClass -> ClassReflection::isAbstract
+   - Building for HTML might throw a missing class error in the reflection cache. 
+     Add missing classes to `artemis.reflect.include` to your gwt.xml.
+
+- Reflection implementation upgraded with latest libGDX reflection improvements.
+- Reflection generator logs a summary of included types per package.
+- **FIX**: Maximum size of reflection cache increased (doubled, more or less).  
+
 #### Version: 2.4.0-SNAPSHOT
 
 - **BREAKING CHANGES**